+++ src/constructs/iam/roles/index.ts
@@ -1,0 +1,2 @@
+export * from "./instance-role";
+export * from "./roles";
